# [go] new key cache backend: "simple"