--- a/go/appencryption/key_cache.go
+++ b/go/appencryption/key_cache.go
@@ -2,6 +2,7 @@ package appencryption
 
 import (
 	"fmt"
+	"strconv"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -130,7 +131,7 @@ func (s *simpleCache) Close() error {
 // cacheKey formats an id and create timestamp to a usable
 // key for storage in a cache.
 func cacheKey(id string, create int64) string {
-	return fmt.Sprintf("%s-%d", id, create)
+	return id + strconv.FormatInt(create, 10)
 }
 
 // keyCacher contains cached keys for reuse.