--- a/go/appencryption/key_cache_benchmark_test.go
+++ b/go/appencryption/key_cache_benchmark_test.go
@@ -33,7 +33,7 @@ func BenchmarkKeyCache_GetOrLoad_MultipleThreadsReadExistingKey(b *testing.B) {
 	c := newKeyCache(CacheTypeIntermediateKeys, NewCryptoPolicy())
 
 	c.keys.Set(cacheKey(testKey, created), cacheEntry{
-		key:      &cachedCryptoKey{CryptoKey: internal.NewCryptoKeyForTest(created, false)},
+		key:      newCachedCryptoKey(internal.NewCryptoKeyForTest(created, false)),
 		loadedAt: time.Now(),
 	})
 
@@ -113,10 +113,6 @@ func BenchmarkKeyCache_GetOrLoad_MultipleThreadsWriteUniqueKeys(b *testing.B) {
 	assert.NotNil(b, c.keys)
 
 	expected := i
-	if expected > DefaultKeyCacheMaxSize {
-		expected = DefaultKeyCacheMaxSize
-	}
-
 	assert.Equal(b, expected, int64(c.keys.Len()))
 }
 
@@ -131,7 +127,7 @@ func BenchmarkKeyCache_GetOrLoad_MultipleThreadsReadRevokedKey(b *testing.B) {
 	assert.NoError(b, err)
 
 	cacheEntry := cacheEntry{
-		key:      &cachedCryptoKey{CryptoKey: key},
+		key:      newCachedCryptoKey(key),
 		loadedAt: time.Unix(created, 0),
 	}
 
@@ -168,7 +164,7 @@ func BenchmarkKeyCache_GetOrLoad_MultipleThreadsRead_NeedReloadKey(b *testing.B)
 	assert.NoError(b, err)
 
 	cacheEntry := cacheEntry{
-		key:      &cachedCryptoKey{CryptoKey: key},
+		key:      newCachedCryptoKey(key),
 		loadedAt: time.Unix(created, 0),
 	}
 
@@ -204,7 +200,7 @@ func BenchmarkKeyCache_GetOrLoad_MultipleThreadsReadUniqueKeys(b *testing.B) {
 
 		c.mapLatestKeyMeta(meta.ID, meta)
 		c.keys.Set(cacheKey(meta.ID, meta.Created), cacheEntry{
-			key:      &cachedCryptoKey{CryptoKey: internal.NewCryptoKeyForTest(created, false), refs: 1},
+			key:      newCachedCryptoKey(internal.NewCryptoKeyForTest(created, false)),
 			loadedAt: time.Now(),
 		})
 	}
@@ -233,7 +229,7 @@ func BenchmarkKeyCache_GetOrLoadLatest_MultipleThreadsReadExistingKey(b *testing
 
 	c.mapLatestKeyMeta(testKey, KeyMeta{testKey, created})
 	c.keys.Set(cacheKey(testKey, created), cacheEntry{
-		key:      &cachedCryptoKey{CryptoKey: internal.NewCryptoKeyForTest(created, false)},
+		key:      newCachedCryptoKey(internal.NewCryptoKeyForTest(created, false)),
 		loadedAt: time.Now(),
 	})
 
@@ -281,6 +277,7 @@ func BenchmarkKeyCache_GetOrLoadLatest_MultipleThreadsWriteUniqueKey(b *testing.
 		for pb.Next() {
 			curr := atomic.AddInt64(&i, 1) - 1
 			_, err := c.GetOrLoadLatest(cacheKey(testKey, curr), func(_ KeyMeta) (*internal.CryptoKey, error) {
+				time.Sleep(5 * time.Millisecond)
 				return internal.NewCryptoKeyForTest(created, false), nil
 			})
 			assert.NoError(b, err)
@@ -289,10 +286,6 @@ func BenchmarkKeyCache_GetOrLoadLatest_MultipleThreadsWriteUniqueKey(b *testing.
 	assert.NotNil(b, c.keys)
 
 	expected := i
-	if expected > DefaultKeyCacheMaxSize {
-		expected = DefaultKeyCacheMaxSize
-	}
-
 	assert.Equal(b, expected, int64(c.keys.Len()))
 }
 
@@ -306,7 +299,7 @@ func BenchmarkKeyCache_GetOrLoadLatest_MultipleThreadsReadStaleRevokedKey(b *tes
 
 	key, err := internal.NewCryptoKey(secretFactory, created, false, []byte("testing"))
 	cacheEntry := cacheEntry{
-		key:      &cachedCryptoKey{CryptoKey: key, refs: 1},
+		key:      newCachedCryptoKey(key),
 		loadedAt: time.Unix(created, 0),
 	}
 
@@ -342,7 +335,7 @@ func BenchmarkKeyCache_GetOrLoadLatest_MultipleThreadsReadRevokedKey(b *testing.
 
 	key, err := internal.NewCryptoKey(secretFactory, created, true, []byte("testing"))
 	cacheEntry := cacheEntry{
-		key:      &cachedCryptoKey{CryptoKey: key, refs: 1},
+		key:      newCachedCryptoKey(key),
 		loadedAt: time.Unix(created, 0),
 	}
 
@@ -384,7 +377,7 @@ func BenchmarkKeyCache_GetOrLoadLatest_MultipleThreadsReadUniqueKeys(b *testing.
 		meta := KeyMeta{ID: keyID, Created: created}
 		c.mapLatestKeyMeta(keyID, meta)
 		c.keys.Set(cacheKey(meta.ID, meta.Created), cacheEntry{
-			key:      &cachedCryptoKey{CryptoKey: internal.NewCryptoKeyForTest(created, false)},
+			key:      newCachedCryptoKey(internal.NewCryptoKeyForTest(created, false)),
 			loadedAt: time.Now(),
 		})
 	}