… go/appencryption/.versionfile